packages/core-py/pyproject.toml
# SPDX-License-Identifier: Apache-2.0
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "revar"
version = "0.1.0"
description = "A deterministic, reproducible test environment for AI browser agents"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [{ name = "The revar contributors" }]
keywords = ["browser", "agents", "evaluation", "testing", "playwright", "ai"]
classifiers = [
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "revar-models",
    "httpx>=0.27",
    "playwright>=1.43",
    "jsonschema>=4.21",
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
    "rich>=13.7",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23"]
"browser-use" = ["browser-use>=0.1.40"]
stagehand = []  # Stagehand is invoked via Node subprocess; only requires Node 20+ at runtime
vision = ["openai>=1.30", "pillow>=10.0"]

[project.scripts]
revar = "revar.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["revar"]
include = ["revar/schemas/*.json"]
